fix up internal_rpc module tree
